pyproject.toml
[tool.poetry]
name = "edivando-calcados"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.5"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
django-filter = "^24.2"
drf-nested-routers = "^0.94.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
